Fix some more code duplication
